🚚 rename: issue #4 contentPost 패키지 추가
