+++ src/main/java/com/example/streaming/contentPost/repository/ContentPostJpaRepository.java
@@ -1,4 +1,4 @@
-package com.example.streaming.repository;
+package com.example.streaming.contentPost.repository;
 
 import com.example.streaming.common.entity.ContentPostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
