+++ src/main/java/com/example/streaming/contentPost/repository/ContentPostRepository.java
@@ -1,0 +1,9 @@
+package com.example.streaming.contentPost.repository;
+
+import com.example.streaming.contentPost.model.ContentPost;
+
+public interface ContentPostRepository {
+
+    ContentPost save(ContentPost contentPost);
+
+}
