+++ src/main/java/com/example/streaming/contentPost/service/ContentPostRepositoryImpl.java
@@ -1,9 +1,9 @@
-package com.example.streaming.service;
+package com.example.streaming.contentPost.service;
 
 import com.example.streaming.common.entity.ContentPostEntity;
-import com.example.streaming.model.ContentPost;
-import com.example.streaming.repository.ContentPostJpaRepository;
-import com.example.streaming.repository.ContentPostRepository;
+import com.example.streaming.contentPost.model.ContentPost;
+import com.example.streaming.contentPost.repository.ContentPostJpaRepository;
+import com.example.streaming.contentPost.repository.ContentPostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
