--- src/main/java/com/example/streaming/repository/ContentPostRepository.java
@@ -1,9 +1,0 @@
-package com.example.streaming.repository;
-
-import com.example.streaming.model.ContentPost;
-
-public interface ContentPostRepository {
-
-    ContentPost save(ContentPost contentPost);
-
-}
